Extract a base class `BaseChange` out of `BodyChange`.
